refactor: restructure directory
